Release 3.0.0 (#4)
# Release 3.0.0

## New features
- Provide features even without CSK_IODDInterpreter
- Possibility to create readMessages without IODD information, by setting start-/endByte and unpack format
- Possibility to search and cut relevant part out of created JSON within readMessage
- New FlowConfig block 'OnNewDataAuto' which automatically creates readMessages and optionally checks to power selected port (WARNING: Do NOT mix with manual readMessage setup)
- React on "OnStopFlowConfigProviders" event of FlowConfig modul to stop pulling IO-Link data
- Check and handle different kind of IOLink.SMI APIs
- Function to delete all existing readMessages
- Check if persistent data to load provides all relevant parameters. Otherwise add default values
- Selectable if timers for readMessages should start automatically after parameters were loaded

## Improvements
- Changed event name "OnNewReadMessage_PORT_MESSAGENAME" to "OnNewReadMessage_INSTANCE_PORT_MESSAGENAME" to make it possible to switch between ports during runtime
- Do not start readMessage timers automatically with creation of readMessage
- Use new event 'OnNewRawReadMessage_INSTANCE_PORT_MESSAGENAME" within FlowConfig (only providing data content as first parameter)
- If using FlowConfig, start readMessage timers after 5 seconds
- Add info about port within "OnNewIOLinkPortStatus" event
- Add list of ports related to instances within "getInstancePortMap" function
- Better handling if CSK_IODDInterpreter is not available

## Bugfix
- Issue with CSK_UserManagement support
- Error if sending process data without any selection
- Error in handling IODD data

+++ CSK_Module_MultiIOLinkSMI/scripts/Communication/MultiIOLinkSMI/MultiIOLinkSMI_Model.lua
@@ -16,19 +16,37 @@ multiIOLinkSMI.__index = multiIOLinkSMI
 
 multiIOLinkSMI.styleForUI = 'None' -- Optional parameter to set UI style
 multiIOLinkSMI.version = Engine.getCurrentAppVersion() -- Version of module
-multiIOLinkSMI.timerActive = true -- Status if read Message timers should run
+multiIOLinkSMI.timerActive = false -- Status if read Message timers should run
 
 -- IO-Link SMI handle used by all instances
-if _G.availableAPIs.specific then
+if _G.availableAPIs.specificSGI and _G.availableAPIs.specificSMI then
+  -- WITH extended SMI API
+  multiIOLinkSMI.IOLinkSMIhandle = IOLink.SMI.create('IOLM1')
+elseif _G.availableAPIs.specificSMI then
+  -- WITHOUT extended SMI API
   multiIOLinkSMI.IOLinkSMIhandle = IOLink.SMI.create()
 end
+
 -- Registering function to be called when any port event occurs 
+----------------------------------------------------------------
+local function handleOnNewPortEventWithEventQualifier(port, qualifier, eventCode)
+  CSK_MultiIOLinkSMI.handleOnNewPortEventWithEventQualifier(port, qualifier, eventCode)
+end
+
 local function handleOnNewPortEvent(port, eventType, eventCode)
   CSK_MultiIOLinkSMI.handleOnNewPortEvent(port, eventType, eventCode)
 end
+
 if multiIOLinkSMI.IOLinkSMIhandle then
-  multiIOLinkSMI.IOLinkSMIhandle:register('OnPortEvent', handleOnNewPortEvent)
+  if _G.availableAPIs.specificSGI then
+    -- WITH extended SMI API
+    multiIOLinkSMI.IOLinkSMIhandle:register('OnPortEvent', handleOnNewPortEventWithEventQualifier)
+  elseif _G.availableAPIs.specificSMI then
+    -- WITHOUT extended SMI API
+    multiIOLinkSMI.IOLinkSMIhandle:register('OnPortEvent', handleOnNewPortEvent)
+  end
 end
+----------------------------------------------------------------
 
 --**************************************************************************
 --********************** End Global Scope **********************************
@@ -118,6 +136,7 @@ function multiIOLinkSMI.create(multiIOLinkSMIInstanceNo)
 
   self.status = 'PORT_NOT_ACTIVE' -- Latest port status to show in UI
   self.activeInUi = false -- Check if this instance is currently active in UI
+  self.readMessageMode = 'IODD' -- Mode if new readMessages should use IODD ('IODD') or not ('NO_IODD')
 
   -- Default values for persistent data
   -- If available, following values will be updated from data of CSK_PersistentData module (check CSK_PersistentData module for this)
@@ -126,21 +145,15 @@ function multiIOLinkSMI.create(multiIOLinkSMIInstanceNo)
 
   -- Parameters to be saved permanently if wanted
   self.parameters = {}
-  self.parameters.flowConfigPriority = CSK_FlowConfig ~= nil or false -- Status if FlowConfig should have priority for FlowConfig relevant configurations
-  self.parameters.name = 'Sensor'.. self.multiIOLinkSMIInstanceNoString -- for future use
-  self.parameters.processingFile = 'CSK_MultiIOLinkSMI_Processing'
-  self.parameters.port = '' -- IOLink port used
-  self.parameters.active = false -- Parameter showing if instance is activated or not
-  self.parameters.ioddInfo = nil -- Table containing IODD information
-  self.parameters.deviceIdentification = nil -- Table containing IODD information
-  self.parameters.ioddReadMessages = {} -- Table contatining information about read messages. Each read message has its own IODD Interpreter instance
-  self.parameters.ioddWriteMessages = {} -- Table contatining information about write messages. Each write message has its own IODD Interpreter instance
+  self.parameters = require('Communication/MultiIOLinkSMI/MultiIOLinkSMI_Parameters') -- Load default parameters
 
   -- Parameters to give to the processing script
   self.multiIOLinkSMIProcessingParams = Container.create()
   self.multiIOLinkSMIProcessingParams:add('multiIOLinkSMIInstanceNumber', multiIOLinkSMIInstanceNo, "INT")
-  self.multiIOLinkSMIProcessingParams:add('SMIhandle', multiIOLinkSMI.IOLinkSMIhandle, 'OBJECT')
-  self.multiIOLinkSMIProcessingParams:add('name', self.parameters.name, 'STRING')
+  if multiIOLinkSMI.IOLinkSMIhandle then
+    self.multiIOLinkSMIProcessingParams:add('SMIhandle', multiIOLinkSMI.IOLinkSMIhandle, 'OBJECT')
+  end
+  --self.multiIOLinkSMIProcessingParams:add('name', self.parameters.name, 'STRING') -- future usage
   self.multiIOLinkSMIProcessingParams:add('active', self.parameters.active, 'BOOL')
   self.multiIOLinkSMIProcessingParams:add('port', self.parameters.port, 'STRING')
 
@@ -327,30 +340,53 @@ end
 
 --- Function to create a read message and instance for the message in CSK_Module_IODDInterpreter
 ---@param messageName string Name of message to create
-function multiIOLinkSMI:createIODDReadMessage(messageName)
+---@param noIODD boolean Status if IODD should be used
+function multiIOLinkSMI:createIODDReadMessage(messageName, noIODD)
+  local useIODD = true
+  local tempIODDInstanceId = ''
+  if noIODD then
+    useIODD = false
+  end
+  if self.parameters.ioddInfo then
+    tempIODDInstanceId = self.parameters.ioddInfo.ioddInstanceId .. '_ReadMessage_' .. messageName
+  end
   self.parameters.ioddReadMessages[messageName] = {
     triggerType = 'Periodic',
     triggerValue = 1000,
-    ioddInstanceId = self.parameters.ioddInfo.ioddInstanceId .. '_ReadMessage_' .. messageName
+    searchBegin = '',
+    searchEnd = '',
+    ioddActive = useIODD,
+    processDataStartByte = 0,
+    processDataEndByte = 1,
+    processDataUnpackFormat = '>I1',
+    ioddInstanceId = tempIODDInstanceId
   }
-  CSK_IODDInterpreter.addInstance()
-  CSK_IODDInterpreter.setInstanceName(self.parameters.ioddReadMessages[messageName].ioddInstanceId)
-  CSK_IODDInterpreter.setSelectedIODD(self.parameters.ioddInfo.ioddName)
-  CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddReadMessages[messageName].ioddInstanceId)
-  if self.parameters.ioddInfo.isProcessDataVariable then
-    CSK_IODDInterpreter.changeProcessDataStructureOptionName(
-      self.parameters.ioddInfo.currentCondition
-    )
+  if useIODD then
+    if CSK_IODDInterpreter then
+      CSK_IODDInterpreter.addInstance()
+      CSK_IODDInterpreter.setInstanceName(self.parameters.ioddReadMessages[messageName].ioddInstanceId)
+      CSK_IODDInterpreter.setSelectedIODD(self.parameters.ioddInfo.ioddName)
+      CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddReadMessages[messageName].ioddInstanceId)
+      if self.parameters.ioddInfo.isProcessDataVariable then
+        CSK_IODDInterpreter.changeProcessDataStructureOptionName(
+          self.parameters.ioddInfo.currentCondition
+        )
+      end
+      CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddReadMessages[messageName].ioddInstanceId)
+    else
+      _G.logger:info(nameOfModule .. ": CSK_IODDInterpreter not available.")
+    end
   end
-  CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddReadMessages[messageName].ioddInstanceId)
 end
 
 --- Function to delete an IODD message and instance in CSK_Module_IODDInterpreter
 ---@param messageToDelete string Message to delete
 function multiIOLinkSMI:deleteIODDReadMessage(messageToDelete)
   if self.parameters.ioddReadMessages[messageToDelete] then
-    CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddReadMessages[messageToDelete].ioddInstanceId)
-    CSK_IODDInterpreter.deleteInstance()
+    if CSK_IODDInterpreter then
+      CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddReadMessages[messageToDelete].ioddInstanceId)
+      CSK_IODDInterpreter.deleteInstance()
+    end
     self.parameters.ioddReadMessages[messageToDelete] = nil
     if not self.parameters.ioddReadMessages then
       self.parameters.ioddReadMessages = {}
@@ -365,29 +401,38 @@ end
 --- Function to create a write message and instance for the message in CSK_Module_IODDInterpreter
 ---@param messageName string Name of message to create
 function multiIOLinkSMI:createIODDWriteMessage(messageName)
-  self.parameters.ioddWriteMessages[messageName] = {
-    ioddInstanceId = self.parameters.ioddInfo.ioddInstanceId .. '_WriteMessage_' .. messageName
-  }
-  CSK_IODDInterpreter.addInstance()
-  CSK_IODDInterpreter.setInstanceName(self.parameters.ioddWriteMessages[messageName].ioddInstanceId)
-  CSK_IODDInterpreter.setSelectedIODD(self.parameters.ioddInfo.ioddName)
-  CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddWriteMessages[messageName].ioddInstanceId)
-  if self.parameters.ioddInfo.isProcessDataVariable then
-    CSK_IODDInterpreter.changeProcessDataStructureOptionName(
-      self.parameters.ioddInfo.currentCondition
-    )
+if CSK_IODDInterpreter then
+    self.parameters.ioddWriteMessages[messageName] = {
+      writeMessageEventName = "",
+      ioddInstanceId = self.parameters.ioddInfo.ioddInstanceId .. '_WriteMessage_' .. messageName
+    }
+    CSK_IODDInterpreter.addInstance()
+    CSK_IODDInterpreter.setInstanceName(self.parameters.ioddWriteMessages[messageName].ioddInstanceId)
+    CSK_IODDInterpreter.setSelectedIODD(self.parameters.ioddInfo.ioddName)
+    CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddWriteMessages[messageName].ioddInstanceId)
+    if self.parameters.ioddInfo.isProcessDataVariable then
+      CSK_IODDInterpreter.changeProcessDataStructureOptionName(
+        self.parameters.ioddInfo.currentCondition
+      )
+    end
+    CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddWriteMessages[messageName].ioddInstanceId)
+  else
+    _G.logger:info(nameOfModule .. ": CSK_IODDInterpreter not available.")
   end
-  CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddWriteMessages[messageName].ioddInstanceId)
 end
 
 --- Function to delete an IODD message and instance in CSK_Module_IODDInterpreter
 ---@param messageToDelete string Message to delete
 function multiIOLinkSMI:deleteIODDWriteMessage(messageToDelete)
-  CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddWriteMessages[messageToDelete].ioddInstanceId)
-  CSK_IODDInterpreter.deleteInstance()
-  self.parameters.ioddWriteMessages[messageToDelete] = nil
-  if not self.parameters.ioddWriteMessages then
-    self.parameters.ioddWriteMessages = {}
+  if CSK_IODDInterpreter then
+    CSK_IODDInterpreter.setSelectedInstance(self.parameters.ioddWriteMessages[messageToDelete].ioddInstanceId)
+    CSK_IODDInterpreter.deleteInstance()
+    self.parameters.ioddWriteMessages[messageToDelete] = nil
+    if not self.parameters.ioddWriteMessages then
+      self.parameters.ioddWriteMessages = {}
+    end
+  else
+    _G.logger:info(nameOfModule .. ": CSK_IODDInterpreter not available.")
   end
 end
